Stay on current tab after submit Highlight current tab Replace New Folk field labels with greyed out text in fields Change submit to floating button

+++ attendance.php
@@ -10,14 +10,16 @@ require_once 'ajaxSetup.php';
  */
 class workshopAttendees {
 	
+	private $count = 0;  //tracks attendee count
 	/*
  	 * Processes posted data
 	 */
 	function process_post() {
 
 		$attendance_nonce = $_REQUEST['attendance_nonce'];
+		
 		if ( wp_verify_nonce($attendance_nonce, 'attendance_nonce') ) {
-			if ($_POST['submit']) {
+			if ($_POST['Update']) {
 				global $wpdb;
 
 				$workshop_id = $_POST['workshop'];
@@ -56,42 +58,47 @@ class workshopAttendees {
 					$email = $_POST[ 'email_' . $i ];
 					$phone = $_POST[ 'phone_' . $i ];
 					$notes = $_POST[ 'notes_' . $i ];
+					if ($firstname != "First Name (Required)") //don't insert New Folk data if first name not entered
+					{
+						$data = array();
+						$format = array();
+						
+						//none of these are required so clear if not changed
+						if ($lastname == "Last Name") {$lastname = "";}
+						if ($email == "Email") {$email = "";}
+						if ($phone == "Phone") {$phone = "";}
+				
+						$data['workshopid'] = $workshop_id;
+						array_push($format, "%d");
+						$data['user_id'] = $user_id;
+						array_push($format, "%d");
+						$data['firstname'] = $firstname;
+						array_push($format, "%s");
+						$data['lastname'] = $lastname;
+						array_push($format, "%s");
+						$data['phone'] = $phone;
+						array_push($format, "%s");
+						$data['email'] = $email;
+						array_push($format, "%s");
+						$data['notes'] = $notes;
+						array_push($format, "%s");
 
-					$data = array();
-					$format = array();
-
-					$data['workshopid'] = $workshop_id;
-					array_push($format, "%d");
-					$data['user_id'] = $user_id;
-					array_push($format, "%d");
-					$data['firstname'] = $firstname;
-					array_push($format, "%s");
-					$data['lastname'] = $lastname;
-					array_push($format, "%s");
-					$data['phone'] = $phone;
-					array_push($format, "%s");
-					$data['email'] = $email;
-					array_push($format, "%s");
-					$data['notes'] = $notes;
-					array_push($format, "%s");
-
-
-					if ($attending && !$id && ($user_id || $firstname || $lastname || $email)) {
-						$wpdb->insert( $table_name,
-							$data,
-							$format);
-					} else if ($attending && $id) {
-						$wpdb->update( $table_name,
-							$data,
-							array ( 'ID' => $id),
-							$format);
-					} else if (!$attending && $id) {
-						$sql = $wpdb->prepare("DELETE FROM `$table_name` WHERE id = %d", $id);
-						$wpdb->query( $sql );
+						if ($attending && !$id && ($user_id || $firstname || $lastname || $email)) {
+							$wpdb->insert( $table_name,
+								$data,
+								$format);
+						} else if ($attending && $id) {
+							$wpdb->update( $table_name,
+								$data,
+								array ( 'ID' => $id),
+								$format);
+						} else if (!$attending && $id) {
+							$sql = $wpdb->prepare("DELETE FROM `$table_name` WHERE id = %d", $id);
+							$wpdb->query( $sql );
+						}
 					}
-
 				}
-
+				
 			}
 		}
 	}
@@ -100,47 +107,43 @@ class workshopAttendees {
 	 * Returns users in database rows
 	 * sorted by first_name or display_name
 	 */
-	function recent_users() {
+	function users($type) {
 		global $wpdb;
 
+		$users_name = $wpdb->prefix . "users";
+		$usermeta_name = $wpdb->prefix . "usermeta";
+		$workshop_attendance_name = $wpdb->prefix . "workshop_attendance";
+		$workshops_name = $wpdb->prefix . "workshops";
+
+		$sqlFilter = ""; //$type == "recent"
+		if ($type == "remaining")
+		{
+			$sqlFilter = "NOT";
+		}
+		
 		$sql = $wpdb->prepare("SELECT display_name, u.ID, user_email
-								FROM  `wp_users` u
-								JOIN  `wp_usermeta` m ON u.id = m.user_id AND m.meta_key =  'first_name'
-								JOIN  `wp_usermeta` m2 ON u.id = m2.user_id AND m2.meta_key IN ('wp_capabilities')								
+								FROM  `$users_name` u
+								JOIN  `$usermeta_name` m ON u.id = m.user_id AND m.meta_key =  'first_name'
+								JOIN  `$usermeta_name` m2 ON u.id = m2.user_id AND m2.meta_key IN ('wp_capabilities')								
 								WHERE u.id <> 1 
-								AND u.id IN 
+								AND u.id " . $sqlFilter . " IN 
 								(
 									SELECT DISTINCT a.user_ID
-									FROM `wp_workshop_attendance` a
-									JOIN `wp_workshops` w ON w.id = a.workshopid 
+									FROM `$workshop_attendance_name` a
+									JOIN `$workshops_name` w ON w.id = a.workshopid 
 									AND w.date >= DATE_SUB(CURDATE(), INTERVAL 6 MONTH)
 								)
-								AND m2.meta_value LIKE '%%author%%'
+								AND 
+									(
+										m2.meta_value LIKE '%%author%%'
+										OR m2.meta_value LIKE '%%editor%%'
+										OR m2.meta_value LIKE '%%administrator%%'
+									)
 								ORDER BY COALESCE( NULLIF( m.meta_value,  '' ) , display_name )", $workshop_id);
-		$recent_users = $wpdb->get_results( $sql, ARRAY_A );
-		return $recent_users;
+
+		$users = $wpdb->get_results( $sql, ARRAY_A );
+		return $users;
 	}
-
-	function remaining_users() {
-		global $wpdb;
-
-		$sql = $wpdb->prepare("SELECT display_name, u.ID, user_email
-								FROM  `wp_users` u
-								JOIN  `wp_usermeta` m ON u.id = m.user_id AND m.meta_key =  'first_name'
-								JOIN  `wp_usermeta` m2 ON u.id = m2.user_id AND m2.meta_key IN ('wp_capabilities')								
-								WHERE u.id <> 1 
-								AND u.id NOT IN 
-								(
-									SELECT DISTINCT a.user_ID
-									FROM `wp_workshop_attendance` a
-									JOIN `wp_workshops` w ON w.id = a.workshopid 
-									AND w.date >= DATE_SUB(CURDATE(), INTERVAL 6 MONTH)
-								)
-								AND m2.meta_value LIKE '%%author%%'
-								ORDER BY COALESCE( NULLIF( m.meta_value,  '' ) , display_name )", $workshop_id);
-		$remaining_users = $wpdb->get_results( $sql, ARRAY_A );
-		return $remaining_users;
-	}	
 
 	/* 
 	 * Returns attendees of workshop in database rows
@@ -152,7 +155,6 @@ class workshopAttendees {
 
 		$sql = $wpdb->prepare("SELECT * FROM `$table_name` WHERE workshopid = %d ORDER BY `lastname`", $workshop_id);
 		$rows = $wpdb->get_results( $sql, ARRAY_A );
-
 		return $rows;
 	}
 
@@ -169,7 +171,6 @@ class workshopAttendees {
 				$attendees[$row['email']] = $row;
 			}
 		}
-
 		return $attendees;
 	}
 
@@ -213,10 +214,83 @@ class workshopAttendees {
 		if ($workshop_id) {
 			$attendee_rows = $this->attendee_rows( $workshop_id );
 		}
-
+		
 		$this->render_form($workshop_id, $attendee_rows);
 	}
 
+	/*
+	 * render attendee tab, recent or remaining
+	 */
+	 
+	function render_attendees($type, $attendees, $rendered_emails)
+	{
+?>
+		<table>
+<?php
+		global $count;
+		$users = $this->users($type);
+		foreach ($users as $user) {
+			$count = $count + 1;
+			$name = $user['display_name'];
+			$user_info = get_userdata($user['ID']);
+			$attendance_info = $attendees[$user['user_email']];
+			$checked = '';
+			$class = "absent";
+
+			array_push($rendered_emails, $user['user_email']);
+			
+			if ($attendance_info) {
+				$checked = 'checked = "checked"';
+				$class = "present";
+			}
+			if ($user_info->first_name || $user_info->last_name) {
+				$name = $user_info->first_name . ' ' . $user_info->last_name;
+			}
+
+?>
+			<tr onclick="selectRow(this)" class="<?php echo $class; ?>"><td>
+				<input onclick="checkClicked(this, event)" type="checkbox" name="attending_<?php echo $count; ?>" value="attending" <?php echo $checked; ?> ><?php echo $name; ?></input>
+<?php
+			if ($user_info->user_description) {
+?>
+			<div class="details">
+				<?php echo $user_info->user_description; ?>
+			</div>
+<?php
+			}
+?>
+			</td>
+			<input type="hidden" name="user_id_<?php echo $count; ?>" value="<?php echo $user['ID']; ?>"/>
+<?php
+			if ($user_info->first_name) {
+?>
+				<input type="hidden" name="firstname_<?php echo $count; ?>" value="<?php echo $user_info->first_name; ?>"/>
+<?php
+			}
+			if ($user_info->last_name && strlen($user_info->last_name)) {
+?>
+				<input type="hidden" name="lastname_<?php echo $count; ?>" value="<?php echo $user_info->last_name; ?>"/>
+<?php
+			} else {
+?>
+				<input type="hidden" name="lastname_<?php echo $count; ?>" value="<?php echo $user['display_name']; ?>"/>
+<?php
+			}
+			if ($attendance_info) {
+?>
+				<input type="hidden" name="id_<?php echo $count; ?>" value="<?php echo $attendance_info["id"]; ?>"/>
+<?php
+			}
+?>
+				<input type="hidden" name="email_<?php echo $count; ?>" value="<?php echo $user['user_email']; ?>"/>
+			</tr>
+<?php
+		}
+?>
+		</table>
+<?php
+	}
+	
 	/*
 	 * renders form
 	 */
@@ -225,167 +299,64 @@ class workshopAttendees {
 		$attendees = $this->attendees( $workshop_id, $attendee_rows );
 
 		global $wpdb;
-
+		global $count;
 		$attendance_nonce = wp_create_nonce('attendance_nonce');
 ?>
 
 	<form method="post" action="<?php echo plugins_url( 'attendance.php', __FILE__ );?>">
 		<input type="hidden" name="attendance_nonce" value="<?php echo $attendance_nonce; ?>" />
+		<input id="Update" type="submit" name="Update" value="Update">
 <?php
 		if ($workshop_id) {
 ?>
 		<input type="hidden" name="workshop" value="<?php echo $workshop_id;?>">
 <?php
 		}
-
 ?>
+<br>
+<br>
 <div class="tabs">
 	<ul>
-	<li class="current"><a href="#recent"><h2>Recent</h2></a></li>
-	<li><a href="#remaining"><h2>Remaining</h2></a></li>
-	<li><a href="#newfolks"><h2>New Folks</h2></a></li>
+<?php 
+	if (isset($_POST['current_tab']))	
+	{
+		$current_tab = $_POST['current_tab'];
+	}
+	else
+	{
+		$current_tab = "#recent";
+	}
+?>
+	<input type="hidden" id="current_tab" name="current_tab" value="<?php echo $current_tab;?>" />
+	<li><a href="#recent"><h3>Recent</h3></a></li>
+	<li><a href="#remaining"><h3>Remaining</h3></a></li>
+	<li><a href="#newfolks"><h3>New Folks</h3></a></li>
 	</ul>
 	<div>
 	<div id="recent" class="tab-content">
-	<table>
 <?php
-	$count = 0;
-	$recent_users = $this->recent_users();
 	$rendered_emails = array();
-	foreach ($recent_users as $user) {
-			$count = $count + 1;
-			$name = $user['display_name'];
-			$user_info = get_userdata($user['ID']);
-			$attendance_info = $attendees[$user['user_email']];
-			$checked = '';
-			$class = "absent";
-
-			array_push($rendered_emails, $user['user_email']);
-			
-			if ($attendance_info) {
-				$checked = 'checked = "checked"';
-				$class = "present";
-			}
-			if ($user_info->first_name || $user_info->last_name) {
-				$name = $user_info->first_name . ' ' . $user_info->last_name;
-			}
-
-			
-?>
-			<tr onclick="selectRow(this)" class="<?php echo $class; ?>"><td>
-				<input onclick="checkClicked(this, event)" type="checkbox" name="attending_<?php echo $count; ?>" value="attending" <?php echo $checked; ?> ><?php echo $name; ?></input>
-<?php
-			if ($user_info->user_description) {
-?>
-			<div class="details">
-
-				<?php echo $user_info->user_description; ?>
-			</div>
-<?php
-			}
-?>
-			</td>
-			<input type="hidden" name="user_id_<?php echo $count; ?>" value="<?php echo $user['ID']; ?>"/>
-<?php
-			if ($user_info->first_name) {
-?>
-				<input type="hidden" name="firstname_<?php echo $count; ?>" value="<?php echo $user_info->first_name; ?>"/>
-<?php
-			}
-			if ($user_info->last_name && strlen($user_info->last_name)) {
-?>
-				<input type="hidden" name="lastname_<?php echo $count; ?>" value="<?php echo $user_info->last_name; ?>"/>
-<?php
-			} else {
-?>
-				<input type="hidden" name="lastname_<?php echo $count; ?>" value="<?php echo $user['display_name']; ?>"/>
-<?php
-			}
-			if ($attendance_info) {
-?>
-				<input type="hidden" name="id_<?php echo $count; ?>" value="<?php echo $attendance_info["id"]; ?>"/>
-<?php
-			}
-?>
-				<input type="hidden" name="email_<?php echo $count; ?>" value="<?php echo $user['user_email']; ?>"/>
-
-			</tr>
-<?php
-	}
-?>
-	</table>
+	$this->render_attendees("recent", $attendees, $rendered_emails);
+?>	
 	</div>
 	<div id="remaining" class="tab-content">
-	<table>
 <?php
-	$remaining_users = $this->remaining_users();
-	$rendered_emails = array();
-	foreach ($remaining_users as $user) {
-			$count = $count + 1;
-			$name = $user['display_name'];
-			$user_info = get_userdata($user['ID']);
-			$attendance_info = $attendees[$user['user_email']];
-			$checked = '';
-			$class = "absent";
-
-			array_push($rendered_emails, $user['user_email']);
-			
-			if ($attendance_info) {
-				$checked = 'checked = "checked"';
-				$class = "present";
-			}
-			if ($user_info->first_name || $user_info->last_name) {
-				$name = $user_info->first_name . ' ' . $user_info->last_name;
-			}
-
-			
-?>
-			<tr onclick="selectRow(this)" class="<?php echo $class; ?>"><td>
-				<input onclick="checkClicked(this, event)" type="checkbox" name="attending_<?php echo $count; ?>" value="attending" <?php echo $checked; ?> ><?php echo $name; ?></input>
-<?php
-			if ($user_info->user_description) {
-?>
-			<div class="details">
-
-				<?php echo $user_info->user_description; ?>
-			</div>
-<?php
-			}
-?>
-			</td>
-			<input type="hidden" name="user_id_<?php echo $count; ?>" value="<?php echo $user['ID']; ?>"/>
-<?php
-			if ($user_info->first_name) {
-?>
-				<input type="hidden" name="firstname_<?php echo $count; ?>" value="<?php echo $user_info->first_name; ?>"/>
-<?php
-			}
-			if ($user_info->last_name && strlen($user_info->last_name)) {
-?>
-				<input type="hidden" name="lastname_<?php echo $count; ?>" value="<?php echo $user_info->last_name; ?>"/>
-<?php
-			} else {
-?>
-				<input type="hidden" name="lastname_<?php echo $count; ?>" value="<?php echo $user['display_name']; ?>"/>
-<?php
-			}
-			if ($attendance_info) {
-?>
-				<input type="hidden" name="id_<?php echo $count; ?>" value="<?php echo $attendance_info["id"]; ?>"/>
-<?php
-			}
-?>
-				<input type="hidden" name="email_<?php echo $count; ?>" value="<?php echo $user['user_email']; ?>"/>
-
-			</tr>
-<?php
-	}
-?>
-	</table>
+	$this->render_attendees("remaining", $attendees, $rendered_emails);
+	$count = $count + 1;
+?>	
 	</div>
 	<div id="newfolks" class="tab-content">
+	<dl>
+				<dd><input type="text" name="firstname_<?php echo $count; ?>" title="First Name (Required)"/></dd>
+				<dd><input type="text" name="lastname_<?php echo $count; ?>" title="Last Name"/></dd>
+				<dd><input type="text" name="email_<?php echo $count; ?>" title="Email"/></dd>
+				<dd><input type="text" name="phone_<?php echo $count; ?>" title="Phone"/></dd>
+				<dd>Notes<br><textarea name="notes_<?php echo $count; ?>"></textarea></dd>
+				<input type="hidden" name="attending_<?php echo $count; ?>" value="attending"/>
+	</dl>
 	<table>
 <?php
+	global $count;
 	$class = "present";
 	foreach ($attendee_rows as $attendee) {
 		if (! in_array($attendee['email'], $rendered_emails)) {
@@ -393,7 +364,6 @@ class workshopAttendees {
 ?>
 			<tr onclick="selectRow(this)" class="<?php echo $class; ?>"><td>
 				<input onclick="checkClicked(this, event)" type="checkbox" checked="checked" name="attending_<?php echo $count; ?>" value="attending" <?php echo $checked; ?> ><?php echo $attendee['firstname'] . " " . $attendee['lastname']; ?></input>
-
 <?php
 			if ($attendee['notes']) {
 ?>
@@ -404,16 +374,14 @@ class workshopAttendees {
 <?php
 			}
 ?>
-
-</td>
+				</td>
+				<input type="hidden" name="user_id_<?php echo $count; ?>" value="<?php echo $attendee['user_id']; ?>"/>
 				<input type="hidden" name="firstname_<?php echo $count; ?>" value="<?php echo $attendee['firstname']; ?>"/>
 				<input type="hidden" name="lastname_<?php echo $count; ?>" value="<?php echo $attendee['lastname']; ?>"/>
 				<input type="hidden" name="email_<?php echo $count; ?>" value="<?php echo $attendee['email']; ?>"/>
 				<input type="hidden" name="phone_<?php echo $count; ?>" value="<?php echo $attendee['phone']; ?>"/>
 				<input type="hidden" name="notes_<?php echo $count; ?>" value="<?php echo stripslashes($attendee['notes']); ?>"/>
 				<input type="hidden" name="id_<?php echo $count; ?>" value="<?php echo stripslashes($attendee['id']); ?>"/>
-
-
 			</tr>
 <?php
 		}
@@ -421,34 +389,45 @@ class workshopAttendees {
 	$count = $count + 1;
 ?>
 	</table>
-	<dl>
-				<dt>First Name</dt>
-				<dd><input type="text" name="firstname_<?php echo $count; ?>"/></dd>
-				<dt>Last Name</dt>
-				<dd><input type="text" name="lastname_<?php echo $count; ?>"/></dd>
-				<dt>Email</dt>
-				<dd><input type="text" name="email_<?php echo $count; ?>"/></dd>
-				<dt>Phone</dt>
-				<dd><input type="text" name="phone_<?php echo $count; ?>"/></dd>
-				<dt>Notes</dt>
-				<dd><textarea name="notes_<?php echo $count; ?>"></textarea></dd>
-				<input type="hidden" name="attending_<?php echo $count; ?>" value="attending"/>
-	</dl>
-			<p><input name="submit" type="submit" value="Update Attendance"></p>
 	</div>
 	</div>	
 </div>
 <script type="text/javascript">
 
+$(".tab-content").hide();
+$("<?php echo $current_tab; ?>").fadeIn();
+var selected_tab = $("a[href='<?php echo $current_tab; ?>']").parent();
+selected_tab.addClass("current");
+
 $(document).ready(function(){
- $(".tabs li").click(function() {
-  $(this).parent().parent().find(".tab-content").hide();
-  var selected_tab = $(this).find("a").attr("href");
-  $(selected_tab).fadeIn();
-  $(this).parent().find("li").removeClass('current');
-  $(this).addClass("current");
-   return false;
+	//tab functionality
+	$(".tabs li").click(function() {
+		$(this).parent().parent().find(".tab-content").hide();
+		var selected_tab = $(this).find("a").attr("href");
+		$(selected_tab).fadeIn();
+		$(this).parent().find("li").removeClass('current');
+		$(this).addClass("current");
+		$("input#current_tab").val(selected_tab);
+		return false;
     });
+
+	//grey input functionality
+    var inputs = $('input[type=text]');
+    inputs.each(function(){
+        $(this).val($(this).attr('title')).addClass('unfocused');
+    });
+    inputs.focus(function(){
+        var input = $(this);
+        if(input.val() == input.attr('title')){
+            $(this).removeClass('unfocused').val('');
+        }
+    });
+    inputs.blur(function(){
+        var input = $(this);
+        if(input.val() == ''){ // User has not placed text
+            input.val(input.attr('title')).addClass('unfocused');
+        }
+    }); 	
 });</script>
 </form>
 <?php
@@ -467,6 +446,7 @@ $workshop_id = $_REQUEST['workshop'];
 <html>
 	<head>
 		<script type="text/javascript" src="js/attendance.js"></script>
+		<script type="text/javascript" src="js/formUI.js"></script>
 		<link rel="stylesheet" type="text/css" href="css/attendance.css" >
 		<link rel="stylesheet" type="text/css" href="css/tab.css" />
 		<script src="http://code.jquery.com/jquery-1.9.1.min.js"></script>
